Update README.md
Add `$font->close();` to example in README file.

In the case where you have a script that uses this library to read / parse a lot of fonts at once, it is a good idea to call `$font->close();` in order to free up the system processes. If you don't, you may get a "failed to open dir: Too many open files" exception thrown.