Fix glyph sizes to a multiple of 4 bytes
Before this change when the font loca table was using the short format (glyph size / 2), an odd-numbered glyph byte length would be incorrectly record. To prevent this issue, when writing a font we will padd glyphs to align with a 4-byte boundary.

Note: Official guidance needed. The 4-byte boundary is based on table requirements, as well as being a multiple of 2.